feat: show user list in notification
